+++ editor/editor-common/src/__tests__/visual-regression/basics.ts
@@ -15,6 +15,6 @@ describe('Snapshot Test: Basic Content', () => {
       page,
       `.${EmojiSharedCssClassName.EMOJI_SPRITE}`,
     );
-    await snapshot(page);
+    await snapshot(page, undefined, '.ProseMirror');
   });
 });